requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-msssim>=0.2.1
matplotlib>=3.7.0
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.65.0
Pillow>=9.5.0